- Alteração do Banco - Ajuste no calculo

+++ hotmart-challenge/busca/src/main/java/br/com/hotmart/challenge/model/data/QuantidadeVendida.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class QuantidadeVendida {
 
 	private Long id;
-	private Long quantidade;
+	private Integer quantidade;
 
 }
